feat: add assets folder to store static images - fix: update import paths after moving images to assets folder
